[Feature] Update hours and start alarm

+++ app/src/main/java/com/startogamu/zikobot/player/alarm/AlarmReceiver.java
@@ -13,7 +13,7 @@ public class AlarmReceiver extends BroadcastReceiver {
     @Override
     public void onReceive(Context context, Intent intent) {
         this.context = context;
-        long alarmId = intent.getLongExtra(Constants.Extra.INSTANCE.getALARM_ID(), -1);
+        int alarmId = intent.getIntExtra(Constants.Extra.INSTANCE.getALARM_ID(), -1);
         context.startService(AlarmService.newInstance(context, alarmId));
     }
